build.sh: Add dry-run and workdir options
These will be useful for using within Github actions.

+++ build.sh
@@ -12,19 +12,22 @@ Usage: $0 [OPTIONS]
 Build a Lokomotive Dashboard image.
 
   --branch                      Headlamp branch to use
+  -w,--workdir                  The workdir (if left empty, a temporary one will be used)
+  -n, --dry-run                 Clone Headlamp and apply changes only
   -h, --help                    Display this help and exit
 
 EOF
 }
 
 TOP_DIR=$(realpath "./$(dirname ${BASH_SOURCE[0]})")
-BUILD_DIR="$(mktemp -d)"
+BUILD_DIR=
 REPO="https://github.com/kinvolk/headlamp.git"
 BRANCH=master
 PLUGINS_DIR="$TOP_DIR/plugins"
 PLUGINS=$(ls ${PLUGINS_DIR})
+DRY_RUN=
 
-ARGS=$(getopt -o "b:h" -l "branch:,help" \
+ARGS=$(getopt -o "b:w:nh" -l "branch:,workdir:,dry-run,help" \
   -n "build.sh" -- "$@")
 eval set -- "$ARGS"
 
@@ -34,6 +37,14 @@ while true; do
       BRANCH=$2
       shift 2
       ;;
+    -n|--dry-run)
+      DRY_RUN=true
+      shift
+    ;;
+    -w|--workdir)
+      BUILD_DIR=$(realpath $2)
+      shift 2
+    ;;
     -h|--help)
     usage
     exit 0
@@ -47,6 +58,13 @@ done
 
 declare -A ASSETS
 ASSETS["$TOP_DIR/assets/lokomotive-logo.svg"]="./frontend/src/resources/logo-light.svg"
+
+RM_DIR=true
+
+if [ -z "$BUILD_DIR" ]; then
+    BUILD_DIR="$(mktemp -d)"
+    RM_DIR=false
+fi
 
 mkdir -p $BUILD_DIR
 
@@ -92,6 +110,10 @@ done
 # Return value
 ret=0
 
+if [ -n $DRY_RUN ]; then
+    exit $ret
+fi
+
 echo
 echo "Creating container..."
 echo
@@ -106,7 +128,9 @@ else
   echo "L8e Dashboard container successfully created!"
 fi
 
-rm -rf $BUILD_DIR
+if [ $RM_DIR = "true" ]; then
+    rm -rf $BUILD_DIR
+fi
 
 popd > /dev/null
